Feat/body validation

+++ src/services/article.service.ts
@@ -6,24 +6,38 @@ import {
   TCreateArticleWithTagId,
   TUpdateArticle,
 } from "../types/article.type";
-import { TagRepository } from "../repositories/tag.repository";
+import { TagService } from "./tag.service";
+import { UserService } from "./user.service";
 
 export class ArticleService {
   private articleRepository: ArticleRepository;
-  private tagRepository: TagRepository;
+  private tagService: TagService;
+  private userService: UserService;
 
   constructor() {
     this.articleRepository = new ArticleRepository();
+    this.tagService = new TagService();
+    this.userService = new UserService();
   }
 
   async createArticle(newArticleInfo: TCreateArticle): Promise<Article> {
     const { thumbnail, title, subtitle, contents, userId, path, tags } =
       newArticleInfo;
+
     const foundArticle = await this.getArticleFindByPath(path);
     if (foundArticle) throw new CustomError(400, "Article already exists.");
 
+    const tagIds: string[] = [];
     for (const tag of tags) {
-      await this.tagRepository.createTag(tag);
+      const tagId = await this.tagService.getOneTagIdByTitle(tag);
+      if (typeof tagId === "string") tagIds.push(tagId);
+    }
+
+    const articleWriter =
+      await this.userService.findOneUserByIdWithPassword(userId);
+
+    if (!articleWriter) {
+      throw new CustomError(400, "User for article writer does not exist.");
     }
 
     const newArticleInfoWithTagId: TCreateArticleWithTagId = {
@@ -31,9 +45,9 @@ export class ArticleService {
       title,
       subtitle,
       contents,
-      userId,
+      user: articleWriter,
       path,
-      tags,
+      tags: tagIds,
     };
     return this.articleRepository.createArticle(newArticleInfoWithTagId);
   }
@@ -46,15 +60,29 @@ export class ArticleService {
     return this.articleRepository.getArticleFindByPath(path);
   }
 
-  async updateArticle(articleId: string, updatedInfo: TUpdateArticle) {
+  async updateArticle(
+    articleId: string,
+    userId: string,
+    updatedInfo: TUpdateArticle,
+  ) {
     const foundArticle = await this.articleRepository.getArticleById(articleId);
     if (!foundArticle) throw new CustomError(400, "Original aticle not found.");
+
+    const writerId = foundArticle.user.id;
+    if (writerId !== userId)
+      throw new CustomError(400, "Only the writer can update the article.");
+
     return this.articleRepository.updateArticle(articleId, updatedInfo);
   }
 
-  async removeArticle(articleId: string) {
-    const foundArticle = this.articleRepository.getArticleById(articleId);
+  async removeArticle(articleId: string, userId: string) {
+    const foundArticle = await this.articleRepository.getArticleById(articleId);
     if (!foundArticle) throw new CustomError(400, "Original aticle not found.");
+
+    const writerId = foundArticle.user.id;
+    if (writerId !== userId)
+      throw new CustomError(400, "Only the writer can remove the article.");
+
     return this.articleRepository.removeArticle(articleId);
   }
